change interval type to int

+++ kernel/proc.h
@@ -103,8 +103,8 @@ struct proc {
   struct file *ofile[NOFILE];  // Open files
   struct inode *cwd;           // Current directory
   char name[16];               // Process name (debugging)
-  uint64 alarm_interval;       // 警报间隔
+  int alarm_interval;       // 警报间隔
   uint64 handler;            // alarm 处理函数
   // 自上一次调用到进程的alarm fn经历了多少滴答
-  uint64 interval_from_prev_alarm;
+  int interval_from_prev_alarm;
 };
